Completados las acciones de vehiculo gasto y gasolina. Securizadas frente a otros usuarios. Completada las funcionalidades de almacenaje de informacion. Siguiente paso crear los derivados

+++ src/DS/TxinbometroBundle/Controller/GastoController.php
@@ -3,7 +3,6 @@
 namespace DS\TxinbometroBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
 use DS\TxinbometroBundle\Entity\Gasto;
 use DS\TxinbometroBundle\Form\GastoType;
 
@@ -11,19 +10,18 @@ use DS\TxinbometroBundle\Form\GastoType;
  * Gasto controller.
  *
  */
-class GastoController extends Controller
-{
+class GastoController extends Controller {
+
     /**
      * Lists all Gasto entities.
      *
      */
-    public function indexAction()
-    {
-        $usuario = $this->container->get('security.context')->getToken()->getUser();
-        
+    public function indexAction() {
+        $vehiculo = $this->container->get('security.context')->getToken()->getUser()->getVehiculo();
+
         $em = $this->getDoctrine()->getEntityManager();
 
-        $entities = $em->getRepository('TxinbometroBundle:Gasto')->getAllFrom($usuario->getId());
+        $entities = $em->getRepository('TxinbometroBundle:Gasto')->getAllFrom($vehiculo->getId());
 
         return $this->render('TxinbometroBundle:Gasto:index.html.twig', array(
             'entities' => $entities
@@ -34,8 +32,9 @@ class GastoController extends Controller
      * Finds and displays a Gasto entity.
      *
      */
-    public function showAction($id)
-    {
+    public function showAction($id) {
+        $usuario = $this->container->get('security.context')->getToken()->getUser();
+
         $em = $this->getDoctrine()->getEntityManager();
 
         $entity = $em->getRepository('TxinbometroBundle:Gasto')->find($id);
@@ -44,10 +43,14 @@ class GastoController extends Controller
             throw $this->createNotFoundException('Unable to find Gasto entity.');
         }
 
+        if ($entity->getVehiculo()->getUsuario()->getId() != $usuario->getId()) {
+            throw $this->createNotFoundException('Unable to access Gasolina entity.');
+        }
+
         $deleteForm = $this->createDeleteForm($id);
 
         return $this->render('TxinbometroBundle:Gasto:show.html.twig', array(
-            'entity'      => $entity,
+            'entity' => $entity,
             'delete_form' => $deleteForm->createView(),
         ));
     }
@@ -56,14 +59,13 @@ class GastoController extends Controller
      * Displays a form to create a new Gasto entity.
      *
      */
-    public function newAction()
-    {
+    public function newAction() {
         $entity = new Gasto();
-        $form   = $this->createForm(new GastoType(), $entity);
+        $form = $this->createForm(new GastoType(), $entity);
 
         return $this->render('TxinbometroBundle:Gasto:new.html.twig', array(
             'entity' => $entity,
-            'form'   => $form->createView()
+            'form' => $form->createView()
         ));
     }
 
@@ -71,12 +73,15 @@ class GastoController extends Controller
      * Creates a new Gasto entity.
      *
      */
-    public function createAction()
-    {
-        $entity  = new Gasto();
+    public function createAction() {
+        $vehiculo = $this->container->get('security.context')->getToken()->getUser()->getVehiculo();
+
+        $entity = new Gasto();
         $request = $this->getRequest();
-        $form    = $this->createForm(new GastoType(), $entity);
+        $form = $this->createForm(new GastoType(), $entity);
         $form->bindRequest($request);
+
+        $entity->setVehiculo($vehiculo);
 
         if ($form->isValid()) {
             $em = $this->getDoctrine()->getEntityManager();
@@ -84,12 +89,11 @@ class GastoController extends Controller
             $em->flush();
 
             return $this->redirect($this->generateUrl('txinbometro_gasto_show', array('id' => $entity->getId())));
-            
         }
 
         return $this->render('TxinbometroBundle:Gasto:new.html.twig', array(
             'entity' => $entity,
-            'form'   => $form->createView()
+            'form' => $form->createView()
         ));
     }
 
@@ -97,8 +101,9 @@ class GastoController extends Controller
      * Displays a form to edit an existing Gasto entity.
      *
      */
-    public function editAction($id)
-    {
+    public function editAction($id) {
+        $usuario = $this->container->get('security.context')->getToken()->getUser();
+
         $em = $this->getDoctrine()->getEntityManager();
 
         $entity = $em->getRepository('TxinbometroBundle:Gasto')->find($id);
@@ -106,13 +111,15 @@ class GastoController extends Controller
         if (!$entity) {
             throw $this->createNotFoundException('Unable to find Gasto entity.');
         }
-
+        if ($entity->getVehiculo()->getUsuario()->getId() != $usuario->getId()) {
+            throw $this->createNotFoundException('Unable to access Gasolina entity.');
+        }
         $editForm = $this->createForm(new GastoType(), $entity);
         $deleteForm = $this->createDeleteForm($id);
 
         return $this->render('TxinbometroBundle:Gasto:edit.html.twig', array(
-            'entity'      => $entity,
-            'edit_form'   => $editForm->createView(),
+            'entity' => $entity,
+            'edit_form' => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
         ));
     }
@@ -121,8 +128,9 @@ class GastoController extends Controller
      * Edits an existing Gasto entity.
      *
      */
-    public function updateAction($id)
-    {
+    public function updateAction($id) {
+        $usuario = $this->container->get('security.context')->getToken()->getUser();
+
         $em = $this->getDoctrine()->getEntityManager();
 
         $entity = $em->getRepository('TxinbometroBundle:Gasto')->find($id);
@@ -131,7 +139,11 @@ class GastoController extends Controller
             throw $this->createNotFoundException('Unable to find Gasto entity.');
         }
 
-        $editForm   = $this->createForm(new GastoType(), $entity);
+        if ($entity->getVehiculo()->getUsuario()->getId() != $usuario->getId()) {
+            throw $this->createNotFoundException('Unable to access Gasolina entity.');
+        }
+
+        $editForm = $this->createForm(new GastoType(), $entity);
         $deleteForm = $this->createDeleteForm($id);
 
         $request = $this->getRequest();
@@ -146,8 +158,8 @@ class GastoController extends Controller
         }
 
         return $this->render('TxinbometroBundle:Gasto:edit.html.twig', array(
-            'entity'      => $entity,
-            'edit_form'   => $editForm->createView(),
+            'entity' => $entity,
+            'edit_form' => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
         ));
     }
@@ -156,8 +168,9 @@ class GastoController extends Controller
      * Deletes a Gasto entity.
      *
      */
-    public function deleteAction($id)
-    {
+    public function deleteAction($id) {
+        $usuario = $this->container->get('security.context')->getToken()->getUser();
+
         $form = $this->createDeleteForm($id);
         $request = $this->getRequest();
 
@@ -170,6 +183,9 @@ class GastoController extends Controller
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find Gasto entity.');
             }
+            if ($entity->getVehiculo()->getUsuario()->getId() != $usuario->getId()) {
+                throw $this->createNotFoundException('Unable to access Gasolina entity.');
+            }
 
             $em->remove($entity);
             $em->flush();
@@ -178,11 +194,11 @@ class GastoController extends Controller
         return $this->redirect($this->generateUrl('txinbometro_gasto'));
     }
 
-    private function createDeleteForm($id)
-    {
+    private function createDeleteForm($id) {
         return $this->createFormBuilder(array('id' => $id))
-            ->add('id', 'hidden')
-            ->getForm()
+                ->add('id', 'hidden')
+                ->getForm()
         ;
     }
+
 }
